+++ test/file_cache_test.dart
@@ -1,32 +1,25 @@
 import 'package:test/test.dart';
 
-import '../lib/file_cache.dart';
+import 'package:file_cache/file_cache.dart';
 
 void main() {
-  test('ExpiredSecondsParse', () {
-    final f = FileCache(path: '');
-
-    expect(f.extractSeconds("max-age=300"), 300);
-    expect(f.extractSeconds("private, max-age = 300"), 300);
-  });
-
   test('getJson', () async {
-    FileCache fileCache = await FileCache.fromDefault(path: "cache3");
+    final fileCache = FileCache(path: "cache3");
 
-    bool res = await fileCache.clean();
-    expect(res, true);
+    await fileCache.clean();
 
-    Map map = await fileCache.getJson('http://httpbin.org/cache/600?a=b');
+    Map map =
+        await fileCache.getJson(Uri.parse('https://httpbin.org/cache/600?a=b'));
     expect(map.length, 4);
     expect(map['args']['a'], 'b');
     expect(fileCache.stats.hitFiles, 0);
   });
 
   test('scan', () async {
-    FileCache fileCache =
-        await FileCache.fromDefault(path: "cache3", scan: true);
+    final fileCache = FileCache(path: "cache3");
 
-    Map map = await fileCache.getJson('http://httpbin.org/cache/600?a=b');
+    Map map =
+        await fileCache.getJson(Uri.parse('https://httpbin.org/cache/600?a=b'));
     expect(map.length, 4);
     expect(map['args']['a'], 'b');
     expect(fileCache.stats.hitFiles, 1);